aws_infrastruture/requirements.txt
aws_cdk.aws_dynamodb
aws_cdk.core
cdk-chalice==0.4.0
